Bluetooth programmer using Arduino nano
Bluetooth programmer using Arduino nano.
Plug in bluetooth HC-05 (one with a button) while holding the button, this will put it into programming mode, press the reset button on the arduino and in a few seconds the slow blinking LED on the bluetooth module will begin to start flashing fast. 
Ive added a cable to the nano, Power, ground, TX, RX (TX and RX are swapped at the bluetooth end) as seen in the picture.
I leave the nano powered on then plug in the bluetooth, do this at your own risk.